backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
torch==2.0.1
transformers==4.33.2
PyYAML==6.0.1
